Creates a security group for each namespace
This patch introduces a separation of security groups
for namespace, used to control the ingress to pods.
Also, a SG is created for each service, associated to
its VIP.

Finally, a new annotation is added to maintain the
SG Rule used to allow the access to the VIP from
the external networks (FIP).

Implements task IN-333

Change-Id: Ifd46768b8aba867013e01cd59c8b0f50fe80dfa4
Signed-off-by: Pablo Chacin <pablo@midokura.com>